requirements.txt
fastapi
uvicorn[standard]
tradingview-ta
yfinance
pydantic-settings
